+++ tests/after-hook.spec.ts
@@ -1,34 +1,69 @@
 import chai from 'chai'
-import http from 'chai-http'
+import { After, HookParams, Request, Response, Service } from '../lib'
+import buildFakeRequest from './buildFakeRequest'
 import server from './fixtures/index'
 const expect = chai.expect
-
-chai.use(http)
 
 describe('@After hook', () => {
   after(() => {
     server.close()
   })
 
-  it('should handle the @After metadata hook on GET', (done: Function) => {    
-    chai
-      .request(server)
-      .get('/users/1')
-      .end((err, res) => {
-        expect(res).to.have.status(200)
-        expect(res.body).to.have.property('metadata').with.property('timestamp')
-        done()
+  it('should handle the @After metadata hook on GET', async () => {    
+    class DemoService implements Service {
+      @After(async (hook: HookParams) => {
+        return {
+          ...hook.result,
+          body: {
+            ...hook.result.body,
+            metadata: {
+              timestamp: Date.now()
+            }
+          }
+        }
       })
+      async post(req: Request): Promise<Response> {
+        return { status: 200 }
+      }
+    }
+
+    const res = await new DemoService().post(buildFakeRequest({
+      ctx: {
+        method: 'POST',
+        state: {}
+      }
+    }))
+
+    expect(res.body).to.have.property('metadata').with.property('timestamp')
   })
 
-  it('should not modify the response if @After has no return value', (done: Function) => {    
-    chai
-      .request(server)
-      .get('/comments')
-      .end((err, res) => {
-        expect(res).to.have.status(200)
-        expect(res.body).to.not.have.property('metadata')
-        done()
+  it('should not allow the response body to be directly modified', async () => {    
+    class DemoService implements Service {
+      @After(async (hook: HookParams) => {
+        const res: Response = hook.result
+        try {
+          res.body.metadata = {
+            timestamp: Date.now()
+          }
+        } catch {}
       })
+      async post(req: Request): Promise<Response> {
+        return {
+          status: 200,
+          body: {
+            foo: 'bar'
+          }
+        }
+      }
+    }
+
+    const res = await new DemoService().post(buildFakeRequest({
+      ctx: {
+        method: 'POST',
+        state: {}
+      }
+    }))
+
+    expect(res.body).not.to.have.property('metadata')
   })
 })